+++ src/app/routing.ts
@@ -42,6 +42,7 @@ import { GridSummaryComponent } from './grid-summaries/grid-summaries.sample';
 import { GridPerformanceSampleComponent } from './grid-performance/grid-performance.sample';
 import { GridSelectionComponent } from './grid-selection/grid-selection.sample';
 import { GridToolbarSampleComponent } from './grid-toolbar/grid-toolbar.sample';
+import { GridToolbarCustomSampleComponent } from './grid-toolbar/grid-toolbar-custom.sample';
 import { GridVirtualizationSampleComponent } from './grid-remote-virtualization/grid-remote-virtualization.sample';
 import { ButtonGroupSampleComponent } from './buttonGroup/buttonGroup.sample';
 import { GridColumnGroupsSampleComponent } from './grid-column-groups/grid-column-groups.sample';
@@ -264,6 +265,10 @@ const appRoutes = [
     {
         path: 'gridToolbar',
         component: GridToolbarSampleComponent
+    },
+    {
+        path: 'gridToolbarCustom',
+        component: GridToolbarCustomSampleComponent
     },
     {
         path: 'gridRemoteVirtualization',